+++ x/evm/types/params_test.go
@@ -1,7 +1,6 @@
 package types
 
 import (
-	sdk "github.com/cosmos/cosmos-sdk/types"
 	"testing"
 
 	"github.com/ethereum/go-ethereum/params"
@@ -23,7 +22,7 @@ func TestParamsValidate(t *testing.T) {
 		{"default", DefaultParams(), false},
 		{
 			"valid",
-			NewParams("ara", true, true, DefaultChainConfig(), DefaultMinGasMultiplier, 2929, 1884, 1344),
+			NewParams("ara", true, true, DefaultChainConfig(), 2929, 1884, 1344),
 			false,
 		},
 		{
@@ -46,11 +45,6 @@ func TestParamsValidate(t *testing.T) {
 			},
 			true,
 		},
-		{
-			"invalid min gas multplier",
-			NewParams("ara", true, true, DefaultChainConfig(), sdk.NewDec(2), 2929, 1884, 1344),
-			true,
-		},
 	}
 
 	for _, tc := range testCases {
@@ -65,7 +59,7 @@ func TestParamsValidate(t *testing.T) {
 }
 
 func TestParamsEIPs(t *testing.T) {
-	params := NewParams("ara", true, true, DefaultChainConfig(), DefaultMinGasMultiplier, 2929, 1884, 1344)
+	params := NewParams("ara", true, true, DefaultChainConfig(), 2929, 1884, 1344)
 	actual := params.EIPs()
 
 	require.Equal(t, []int([]int{2929, 1884, 1344}), actual)
@@ -78,7 +72,6 @@ func TestParamsValidatePriv(t *testing.T) {
 	require.NoError(t, validateBool(true))
 	require.Error(t, validateEIPs(""))
 	require.NoError(t, validateEIPs([]int64{1884}))
-	require.Error(t, validateMinGasMultiplier(sdk.NewDec(-5)))
 }
 
 func TestValidateChainConfig(t *testing.T) {