+++ x/feemarket/keeper/eip1559_test.go
@@ -28,14 +28,14 @@ func (suite *KeeperTestSuite) TestCalculateBaseFee() {
 			suite.app.FeeMarketKeeper.GetParams(suite.ctx).BaseFee.BigInt(),
 		},
 		{
-			"with BaseFee - parent block used the same gas as its target",
+			"with BaseFee - parent block wanted the same gas as its target",
 			false,
 			func() {
 				// non initial block
 				suite.ctx = suite.ctx.WithBlockHeight(1)
 
 				// Set gas used
-				suite.app.FeeMarketKeeper.SetBlockGasUsed(suite.ctx, 100)
+				suite.app.FeeMarketKeeper.SetBlockGasWanted(suite.ctx, 100)
 
 				// Set target/gasLimit through Consensus Param MaxGas
 				blockParams := abci.BlockParams{
@@ -53,12 +53,12 @@ func (suite *KeeperTestSuite) TestCalculateBaseFee() {
 			suite.app.FeeMarketKeeper.GetParams(suite.ctx).BaseFee.BigInt(),
 		},
 		{
-			"with BaseFee - parent block used more gas than its target",
+			"with BaseFee - parent block wanted more gas than its target",
 			false,
 			func() {
 				suite.ctx = suite.ctx.WithBlockHeight(1)
 
-				suite.app.FeeMarketKeeper.SetBlockGasUsed(suite.ctx, 200)
+				suite.app.FeeMarketKeeper.SetBlockGasWanted(suite.ctx, 200)
 
 				blockParams := abci.BlockParams{
 					MaxGas:   100,
@@ -74,12 +74,12 @@ func (suite *KeeperTestSuite) TestCalculateBaseFee() {
 			big.NewInt(1125000000),
 		},
 		{
-			"with BaseFee - Parent gas used smaller than parent gas target",
+			"with BaseFee - Parent gas wanted smaller than parent gas target",
 			false,
 			func() {
 				suite.ctx = suite.ctx.WithBlockHeight(1)
 
-				suite.app.FeeMarketKeeper.SetBlockGasUsed(suite.ctx, 50)
+				suite.app.FeeMarketKeeper.SetBlockGasWanted(suite.ctx, 50)
 
 				blockParams := abci.BlockParams{
 					MaxGas:   100,