+++ x/feemarket/types/keys.go
@@ -10,15 +10,28 @@ const (
 
 	// RouterKey uses module name for routing
 	RouterKey = ModuleName
+
+	// TransientKey is the key to access the FeeMarket transient store, that is reset
+	// during the Commit phase.
+	TransientKey = "transient_" + ModuleName
 )
 
 // prefix bytes for the feemarket persistent store
 const (
-	prefixBlockGasUsed      = iota + 1
+	prefixBlockGasWanted    = iota + 1
 	deprecatedPrefixBaseFee // nolint
+)
+
+const (
+	prefixTransientBlockGasUsed = iota + 1
 )
 
 // KVStore key prefixes
 var (
-	KeyPrefixBlockGasUsed = []byte{prefixBlockGasUsed}
+	KeyPrefixBlockGasWanted = []byte{prefixBlockGasWanted}
+)
+
+// Transient Store key prefixes
+var (
+	KeyPrefixTransientBlockGasWanted = []byte{prefixTransientBlockGasUsed}
 )