+++ x/feemarket/types/params.go
@@ -18,6 +18,14 @@ var (
 	ParamStoreKeyBaseFee                  = []byte("BaseFee")
 	ParamStoreKeyEnableHeight             = []byte("EnableHeight")
 	ParamStoreKeyMinGasPrice              = []byte("MinGasPrice")
+	ParamStoreKeyMinGasMultiplier         = []byte("MinGasMultiplier")
+)
+
+var (
+	// DefaultMinGasMultiplier is 0.5 or 50%
+	DefaultMinGasMultiplier = sdk.NewDecWithPrec(50, 2)
+	// DefaultMinGasPrice is 0 (i.e disabled)
+	DefaultMinGasPrice = sdk.ZeroDec()
 )
 
 // ParamKeyTable returns the parameter key table.
@@ -33,6 +41,7 @@ func NewParams(
 	baseFee uint64,
 	enableHeight int64,
 	minGasPrice sdk.Dec,
+	minGasPriceMultiplier sdk.Dec,
 ) Params {
 	return Params{
 		NoBaseFee:                noBaseFee,
@@ -41,6 +50,7 @@ func NewParams(
 		BaseFee:                  sdk.NewIntFromUint64(baseFee),
 		EnableHeight:             enableHeight,
 		MinGasPrice:              minGasPrice,
+		MinGasMultiplier:         minGasPriceMultiplier,
 	}
 }
 
@@ -52,7 +62,8 @@ func DefaultParams() Params {
 		ElasticityMultiplier:     params.ElasticityMultiplier,
 		BaseFee:                  sdk.NewIntFromUint64(params.InitialBaseFee),
 		EnableHeight:             0,
-		MinGasPrice:              sdk.ZeroDec(),
+		MinGasPrice:              DefaultMinGasPrice,
+		MinGasMultiplier:         DefaultMinGasMultiplier,
 	}
 }
 
@@ -65,6 +76,7 @@ func (p *Params) ParamSetPairs() paramtypes.ParamSetPairs {
 		paramtypes.NewParamSetPair(ParamStoreKeyBaseFee, &p.BaseFee, validateBaseFee),
 		paramtypes.NewParamSetPair(ParamStoreKeyEnableHeight, &p.EnableHeight, validateEnableHeight),
 		paramtypes.NewParamSetPair(ParamStoreKeyMinGasPrice, &p.MinGasPrice, validateMinGasPrice),
+		paramtypes.NewParamSetPair(ParamStoreKeyMinGasMultiplier, &p.MinGasMultiplier, validateMinGasPrice),
 	}
 }
 
@@ -80,6 +92,10 @@ func (p Params) Validate() error {
 
 	if p.EnableHeight < 0 {
 		return fmt.Errorf("enable height cannot be negative: %d", p.EnableHeight)
+	}
+
+	if err := validateMinGasMultiplier(p.MinGasMultiplier); err != nil {
+		return err
 	}
 
 	return validateMinGasPrice(p.MinGasPrice)
@@ -159,5 +175,26 @@ func validateMinGasPrice(i interface{}) error {
 		return fmt.Errorf("value cannot be negative: %s", i)
 	}
 
+	return nil
+}
+
+func validateMinGasMultiplier(i interface{}) error {
+	v, ok := i.(sdk.Dec)
+
+	if !ok {
+		return fmt.Errorf("invalid parameter type: %T", i)
+	}
+
+	if v.IsNil() {
+		return fmt.Errorf("invalid parameter: nil")
+	}
+
+	if v.IsZero() || v.IsNegative() {
+		return fmt.Errorf("value cannot be zero or negative: %T", i)
+	}
+
+	if v.GT(sdk.OneDec()) {
+		return fmt.Errorf("value cannot be greater than 1: %T", i)
+	}
 	return nil
 }